Removing the sqldb.Conn interface.
Now using *mysql.Conn everywhere.
